+++ google-cloud/container-linux/kubernetes/apiserver.tf
@@ -1,15 +1,15 @@
 # TCP Proxy load balancer DNS record
 resource "google_dns_record_set" "apiserver" {
   # DNS Zone name where record should be created
-  managed_zone = "${var.dns_zone_name}"
+  managed_zone = var.dns_zone_name
 
   # DNS record
-  name = "${format("%s.%s.", var.cluster_name, var.dns_zone)}"
+  name = format("%s.%s.", var.cluster_name, var.dns_zone)
   type = "A"
   ttl  = 300
 
   # IPv4 address of apiserver TCP Proxy load balancer
-  rrdatas = ["${google_compute_global_address.apiserver-ipv4.address}"]
+  rrdatas = [google_compute_global_address.apiserver-ipv4.address]
 }
 
 # Static IPv4 address for the TCP Proxy Load Balancer
@@ -21,17 +21,17 @@ resource "google_compute_global_address" "apiserver-ipv4" {
 # Forward IPv4 TCP traffic to the TCP proxy load balancer
 resource "google_compute_global_forwarding_rule" "apiserver" {
   name        = "${var.cluster_name}-apiserver"
-  ip_address  = "${google_compute_global_address.apiserver-ipv4.address}"
+  ip_address  = google_compute_global_address.apiserver-ipv4.address
   ip_protocol = "TCP"
   port_range  = "443"
-  target      = "${google_compute_target_tcp_proxy.apiserver.self_link}"
+  target      = google_compute_target_tcp_proxy.apiserver.self_link
 }
 
 # Global TCP Proxy Load Balancer for apiservers
 resource "google_compute_target_tcp_proxy" "apiserver" {
   name            = "${var.cluster_name}-apiserver"
   description     = "Distribute TCP load across ${var.cluster_name} controllers"
-  backend_service = "${google_compute_backend_service.apiserver.self_link}"
+  backend_service = google_compute_backend_service.apiserver.self_link
 }
 
 # Global backend service backed by unmanaged instance groups
@@ -46,26 +46,30 @@ resource "google_compute_backend_service" "apiserver" {
 
   # controller(s) spread across zonal instance groups
   backend {
-    group = "${google_compute_instance_group.controllers.0.self_link}"
+    group = google_compute_instance_group.controllers[0].self_link
   }
 
   backend {
-    group = "${google_compute_instance_group.controllers.1.self_link}"
+    group = google_compute_instance_group.controllers[1].self_link
   }
 
   backend {
-    group = "${google_compute_instance_group.controllers.2.self_link}"
+    group = google_compute_instance_group.controllers[2].self_link
   }
 
-  health_checks = ["${google_compute_health_check.apiserver.self_link}"]
+  health_checks = [google_compute_health_check.apiserver.self_link]
 }
 
 # Instance group of heterogeneous (unmanged) controller instances
 resource "google_compute_instance_group" "controllers" {
-  count = "${length(local.zones)}"
+  count = length(local.zones)
 
-  name = "${format("%s-controllers-%s", var.cluster_name, element(local.zones, count.index))}"
-  zone = "${element(local.zones, count.index)}"
+  name = format(
+    "%s-controllers-%s",
+    var.cluster_name,
+    element(local.zones, count.index),
+  )
+  zone = element(local.zones, count.index)
 
   named_port {
     name = "apiserver"
@@ -73,11 +77,11 @@ resource "google_compute_instance_group" "controllers" {
   }
 
   # add instances in the zone into the instance group
-  instances = [
-    "${matchkeys(google_compute_instance.controllers.*.self_link,
-      google_compute_instance.controllers.*.zone,
-      list(element(local.zones, count.index)))}",
-  ]
+  instances = matchkeys(
+    google_compute_instance.controllers.*.self_link,
+    google_compute_instance.controllers.*.zone,
+    [element(local.zones, count.index)],
+  )
 }
 
 # TCP health check for apiserver
@@ -95,3 +99,4 @@ resource "google_compute_health_check" "apiserver" {
     port = "443"
   }
 }
+