+++ google-cloud/container-linux/kubernetes/workers/outputs.tf
@@ -2,12 +2,13 @@
 
 output "instance_group" {
   description = "Worker managed instance group full URL"
-  value       = "${google_compute_region_instance_group_manager.workers.instance_group}"
+  value       = google_compute_region_instance_group_manager.workers.instance_group
 }
 
 # Outputs for regional load balancing
 
 output "target_pool" {
   description = "Worker target pool self link"
-  value       = "${google_compute_target_pool.workers.self_link}"
+  value       = google_compute_target_pool.workers.self_link
 }
+