+++ google-cloud/container-linux/kubernetes/workers/target_pool.tf
@@ -1,11 +1,11 @@
 # Target pool for TCP/UDP load balancing
 resource "google_compute_target_pool" "workers" {
   name             = "${var.name}-worker-pool"
-  region           = "${var.region}"
+  region           = var.region
   session_affinity = "NONE"
 
   health_checks = [
-    "${google_compute_http_health_check.workers.name}",
+    google_compute_http_health_check.workers.name,
   ]
 }
 
@@ -20,3 +20,4 @@ resource "google_compute_http_health_check" "workers" {
   port         = 10254
   request_path = "/healthz"
 }
+